requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
termcolor==2.4.0
pytest==7.4.3